refactor: extract parameter store logic into useParameterStore hook
- Removed inline parameterStore setup from App.tsx.
- Introduced dedicated useParameterStore hook to encapsulate:
  - Initial fetch of all parameters from backend.
  - Subscription to parameter.update socket events.
  - Automatic cleanup on unmount.
- Simplifies App.tsx and centralises parameter store logic.
